Update apps with new kernel files from new op2.m

+++ apps/c/jac2/res.h
@@ -1,3 +1,3 @@
-inline void res(double *A, double *u, double *du, const double *beta){
+inline void res(double *A, float *u, float *du, const float *beta){
   *du += (*beta)*(*A)*(*u);
 }
